Moving hygrometric conversions to a standalone module
