+++ tests/hygro/test_hygro.py
@@ -23,8 +23,8 @@ PATM = np.linspace(60, 110, 31)
     argnames="formula", argvalues=["Allen1998", "Alduchov1996", "Sonntag1990"]
 )
 def test_calc_vp_sat(bigleaf, formula):
+    from pyrealm.hygro import calc_vp_sat
     from pyrealm.param_classes import HygroParams
-    from pyrealm.utilities import calc_vp_sat
 
     params = HygroParams(magnus_option=formula)
     results = calc_vp_sat(TEMP, params)
@@ -37,8 +37,8 @@ def test_calc_vp_sat(bigleaf, formula):
     argnames="formula", argvalues=["Allen1998", "Alduchov1996", "Sonntag1990"]
 )
 def test_convert_vp_to_vpd(bigleaf, formula):
+    from pyrealm.hygro import convert_vp_to_vpd
     from pyrealm.param_classes import HygroParams
-    from pyrealm.utilities import convert_vp_to_vpd
 
     params = HygroParams(magnus_option=formula)
     results = convert_vp_to_vpd(VP, TEMP, params)
@@ -51,8 +51,8 @@ def test_convert_vp_to_vpd(bigleaf, formula):
     argnames="formula", argvalues=["Allen1998", "Alduchov1996", "Sonntag1990"]
 )
 def test_convert_rh_to_vpd(bigleaf, formula):
+    from pyrealm.hygro import convert_rh_to_vpd
     from pyrealm.param_classes import HygroParams
-    from pyrealm.utilities import convert_rh_to_vpd
 
     params = HygroParams(magnus_option=formula)
     results = convert_rh_to_vpd(RH, TEMP, params)
@@ -62,7 +62,7 @@ def test_convert_rh_to_vpd(bigleaf, formula):
 
 
 def test_convert_sh_to_vp(bigleaf):
-    from pyrealm.utilities import convert_sh_to_vp
+    from pyrealm.hygro import convert_sh_to_vp
 
     results = convert_sh_to_vp(SH, PATM)
     expected = np.array(bigleaf["convert_sh_to_vp"])
@@ -74,8 +74,8 @@ def test_convert_sh_to_vp(bigleaf):
     argnames="formula", argvalues=["Allen1998", "Alduchov1996", "Sonntag1990"]
 )
 def test_convert_sh_to_vpd(bigleaf, formula):
+    from pyrealm.hygro import convert_sh_to_vpd
     from pyrealm.param_classes import HygroParams
-    from pyrealm.utilities import convert_sh_to_vpd
 
     params = HygroParams(magnus_option=formula)
     results = convert_sh_to_vpd(SH, TEMP, PATM, params)
